MC 1.17.10

+++ src/MiNET/MiNET/Net/McpeAvailableCommands.cs
@@ -123,7 +123,7 @@ namespace MiNET.Net
 					command.Versions = new Version[1];
 					string commandName = ReadString();
 					string description = ReadString();
-					int flags = ReadByte();
+					int flags = ReadShort();
 					int permissions = ReadByte();
 
 					command.Name = commandName;
@@ -424,7 +424,7 @@ namespace MiNET.Net
 				{
 					Write(command.Name);
 					Write(command.Versions[0].Description);
-					Write((byte) 0); // flags
+					Write((short) 0); // flags
 					Write((byte) command.Versions[0].CommandPermission); // permissions
 
 					if (command.Versions[0].Aliases.Length > 0)
